requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.11.0
anyio==4.12.0
asttokens==3.0.1
audioop-lts==0.2.2
azure-ai-documentintelligence==1.0.2
azure-core==1.37.0
azure-identity==1.25.1
beautifulsoup4==4.14.3
cachetools==6.2.4
calcpy @ https://github.com/idanpa/calcpy/archive/main.zip#sha256=1e9208073fb9943095cad78a5fad857e296215f8304e071bfbf27942a4b07988
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dateparser==1.2.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
kiwisolver==1.4.9
latex2sympy2==1.9.1
lxml==6.0.2
magika==0.6.3
mammoth==1.11.0
markdownify==1.2.2
-e git+https://github.com/microsoft/markitdown@dde250a456d178fe344fce17ef10d00fe929f680#egg=markitdown&subdirectory=packages/markitdown
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
numpy==2.4.0
olefile==0.47
onnxruntime==1.23.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20251230
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.0.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SpeechRecognition==3.14.4
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.1
traitlets==5.14.3
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
youtube-transcript-api==1.0.3
